+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.0.4
+
+- replace outline effect with superior one
+- add find head button
+- add way to create single sticker
+- fix infinite loop by hardcoding sticker count
+- form to override parameters
+- animation delay from 2sec to 1sec
+- fixed head offset not applying
+- added confirmation button
+
 # 0.0.3
 
 - fix ImageMagick batch script not resizing to required 512px